+++ txtracker/chains/blockdaemon.go
@@ -1,4 +1,4 @@
-package connectors
+package chains
 
 import (
 	"context"
