+++ txtracker/chains/tx.go
@@ -1,4 +1,4 @@
-package connectors
+package chains
 
 import (
 	"context"
